+++ apps/@sparrow-desktop/src/pages/workspace-page/CollectionPage.ViewModel.ts
@@ -2032,7 +2032,6 @@ export default class CollectionsViewModel {
     folder: CollectionItemsDto,
     request: CollectionItemsDto,
   ) => {
-    console.log("opening request", request);
     const requestTabAdapter = new RequestTabAdapter();
     const adaptedRequest = requestTabAdapter.adapt(
       workspaceId || "",
@@ -2055,7 +2054,6 @@ export default class CollectionsViewModel {
     folder: CollectionItemsDto,
     _graphql: CollectionItemsDto,
   ) => {
-    console.log("opening graph ql");
     const requestTabAdapter = new GraphqlTabAdapter();
     const adaptedRequest = requestTabAdapter.adapt(
       workspaceId || "",
