+++ packages/@sparrow-common/src/components/global-search/sub-components/RecentItems.svelte
@@ -140,7 +140,6 @@
       filteredWorkspaces = filteredWorkspaces.map(
         (workspace) => workspace._data,
       );
-      console.log("filteredWorkspaces", filteredWorkspaces);
 
       const recentTestflows = await recentTestflow();
       filteredTestflows = recentTestflows.map((testflow) => testflow._data);
